requirements.txt
dnspython
requests
configparser

